Render Delete Button on Current User's Uploads
The problem was that the delete button was still rendering on an image
after sign out. Instead of keeping the editable property on the uploads
model in the backend, we used a computed property on the front end using
ember promise helpers
(https://github.com/fivetanley/ember-promise-helpers)

The computed property is checking equality of the upload.user.id and the
auth.credentials.id (by injecting the auth service). PROTIP: if ids look
equal and you're getting false for equality check the type of to ensure
the primitive values match. Upload user id was coming through as a
string and auth id was a integer, so we parsed the user id to be
an integer.

In the template we've added the ember helper (await)
to the if statement on the delete button.